+++ src/arche/readers/items.py
@@ -1,4 +1,4 @@
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 import numbers
 from typing import Optional
 
@@ -9,7 +9,41 @@ from scrapinghub import ScrapinghubClient
 from scrapinghub.client.jobs import Job
 
 
-class Items(ABC):
+class Items:
+    def __init__(self, df: pd.DataFrame, expand: bool = False):
+        self.df = self.process_df(df)
+        self._flat_df = None
+        self.expand = expand
+
+    def __len__(self):
+        return len(self.df)
+
+    @property
+    def flat_df(self):
+        if self._flat_df is None:
+            if self.expand:
+                self._flat_df, self._columns_map = pandas.flatten_df(self.df)
+            else:
+                self._flat_df = self.df
+                self._columns_map = {}
+        return self._flat_df
+
+    def process_df(self, df: pd.DataFrame) -> pd.DataFrame:
+        # clean empty objects - mainly lists and dicts, but keep everything else
+        return df.applymap(lambda x: x if x or isinstance(x, numbers.Real) else None)
+
+    def get_origin_column_name(self, column_name: str) -> str:
+        return self._columns_map.get(column_name, column_name)
+
+    @classmethod
+    def from_df(cls, df: pd.DataFrame, expand: bool = True):
+        if "_key" not in df.columns:
+            df["_key"] = df.index
+            df["_key"] = df["_key"].apply(str)
+        return cls(df, expand)
+
+
+class CloudItems(Items):
     def __init__(
         self,
         key: str,
@@ -21,9 +55,10 @@ class Items(ABC):
         self._count = count
         self._limit = None
         self.filters = filters
-        self._df = None
-        self._flat_df = None
         self.expand = expand
+        df = self.fetch_data()
+        df["_key"] = self.format_keys(df["_key"])
+        super().__init__(df=df, expand=expand)
 
     @property
     @abstractmethod
@@ -41,44 +76,22 @@ class Items(ABC):
     def fetch_data(self):
         raise NotImplementedError
 
-    @property
-    def df(self):
-        if self._df is None:
-            self._df = self.process_df(self.fetch_data())
-        return self._df
-
-    @property
-    def size(self):
-        return len(self.df)
-
-    @property
-    def flat_df(self):
-        if self._flat_df is None:
-            if self.expand:
-                self._flat_df, self._columns_map = pandas.flatten_df(self.df)
-            else:
-                self._flat_df = self.df
-                self._columns_map = {}
-        return self._flat_df
-
-    def process_df(self, df: pd.DataFrame) -> pd.DataFrame:
-        # clean empty objects - mainly lists and dicts, but keep everything else
-        df = df.applymap(lambda x: x if x or isinstance(x, numbers.Real) else None)
-        df["_key"] = self.format_keys(df["_key"])
-        return df
-
     def format_keys(self, keys: pd.Series) -> pd.Series:
         raise NotImplementedError
 
-    def get_origin_column_name(self, column_name: str) -> str:
-        return self._columns_map.get(column_name, column_name)
 
-
-class JobItems(Items):
-    def __init__(self, start: int = 0, *args, **kwargs):
-        self.start_index = start
+class JobItems(CloudItems):
+    def __init__(
+        self,
+        key: str,
+        start: int = 0,
+        count: Optional[int] = None,
+        filters: Optional[api.Filters] = None,
+        expand: bool = True,
+    ):
+        self.start_index: int = start
         self._job: Job = None
-        super().__init__(*args, **kwargs)
+        super().__init__(key, count, filters, expand)
 
     @property
     def limit(self) -> int:
@@ -115,7 +128,7 @@ class JobItems(Items):
         )
 
 
-class CollectionItems(Items):
+class CollectionItems(CloudItems):
     @property
     def limit(self) -> int:
         if not self._limit: