+++ src/arche/tools/schema.py
@@ -2,24 +2,44 @@ from dataclasses import dataclass
 import json
 import pprint
 import random
-from typing import List, Optional
+from typing import Any, Dict, List, Optional
 
 from arche.readers.schema import Schema
 from arche.schema_definitions import extension
 from arche.tools import api, helpers
 from genson import SchemaBuilder
-from scrapinghub import ScrapinghubClient
+
+
+@dataclass
+class BasicSchema:
+    d: Schema
+
+    def json(self):
+        print(json.dumps(self.d, indent=4))
+
+    def __repr__(self):
+        return pprint.pformat(self.d)
+
+
+def basic_json_schema(data_source: str, items_numbers: List[int] = None) -> BasicSchema:
+    """Print a json schema based on the provided job_key and item numbers
+
+    Args:
+        data_source: a collection or job key
+        items_numbers: array of item numbers to create schema from
+    """
+    schema = create_json_schema(data_source, items_numbers)
+    return BasicSchema(schema)
 
 
 def create_json_schema(
     source_key: str, item_numbers: Optional[List[int]] = None
-) -> dict:
+) -> Schema:
     if helpers.is_collection_key(source_key):
         store = api.get_collection(source_key)
         items_count = store.count()
     elif helpers.is_job_key(source_key):
-        job = ScrapinghubClient().get_job(source_key)
-        items_count = api.get_items_count(job)
+        items_count = api.get_items_count(api.get_job(source_key))
     else:
         raise ValueError(f"'{source_key}' is not a valid job or collection key")
 
@@ -37,12 +57,12 @@ def create_json_schema(
     samples = []
     for n in item_numbers:
         items = api.get_items(source_key, start_index=n, count=1, p_bar=None)
-        samples.append(items[0])
+        samples.append(items.iloc[0].to_dict())
 
     return infer_schema(samples)
 
 
-def infer_schema(samples):
+def infer_schema(samples: List[Dict[str, Any]]) -> Schema:
     builder = SchemaBuilder("http://json-schema.org/draft-07/schema#")
     for s in samples:
         builder.add_object(s)
@@ -51,7 +71,7 @@ def infer_schema(samples):
     return builder.to_schema()
 
 
-def set_item_no(items_count):
+def set_item_no(items_count: int) -> List[int]:
     """Generate random numbers within items_count range
 
     Returns:
@@ -60,25 +80,3 @@ def set_item_no(items_count):
     if items_count <= 4:
         return [i for i in range(items_count)]
     return random.sample(range(0, items_count), 4)
-
-
-def basic_json_schema(data_source: str, items_numbers: List[int] = None):
-    """Prints a json schema based on the provided job_key and item numbers
-
-    Args:
-        data_source: a collection or job key
-        items_numbers: array of item numbers to create schema from
-    """
-    schema = create_json_schema(data_source, items_numbers)
-    return BasicSchema(schema)
-
-
-@dataclass
-class BasicSchema:
-    d: Schema
-
-    def json(self):
-        print(json.dumps(self.d, indent=4))
-
-    def __repr__(self):
-        return pprint.pformat(self.d)