+++ tests/tools/test_schema_tools.py
@@ -42,6 +42,31 @@ def test_basic_json_schema(mocker):
     mocked_create_js.assert_called_once_with("235801/1/15", [0, 5])
 
 
+def test_create_json_schema(mocker, get_job, get_items):
+    mocker.patch("arche.tools.api.get_job", return_value=get_job, autospec=True)
+    mocker.patch("arche.tools.api.get_items", return_value=get_items, autospec=True)
+    assert schema_tools.create_json_schema(get_job.key, [2, 3]) == {
+        "$schema": "http://json-schema.org/draft-07/schema#",
+        "definitions": {
+            "float": {"pattern": "^-?[0-9]+\\.[0-9]{2}$"},
+            "url": {
+                "pattern": (
+                    "^https?://(www\\.)?[a-z0-9.-]*\\.[a-z]{2,}"
+                    "([^<>%\\x20\\x00-\\x1f\\x7F]|%[0-9a-fA-F]{2})*$"
+                )
+            },
+        },
+        "additionalProperties": False,
+        "type": "object",
+        "properties": {
+            "_key": {"type": "string"},
+            "_type": {"type": "string"},
+            "name": {"type": "string"},
+        },
+        "required": ["_key", "_type", "name"],
+    }
+
+
 def test_basic_schema_json(capsys):
     bs = schema_tools.BasicSchema(
         {