+++ CHANGES.md
@@ -12,10 +12,12 @@ Note that the top-most release is changes in the unreleased master branch on Git
 
 ## [0.4.0] (Work In Progress)
 ### Added
+- Accept dataframes as source or target, #69
 ### Changed
 - data_quality_report plots the same "Fields Coverage" instead of green "Scraped Fields Coverage"
 - Plot theme changed from ggplot2 to seaborn, #62
 - Same target and source raise an error
+- Brought back error colours in `report_all()`. Also passed rules marked with green PASSED.
 ### Fixed
 - Online documentation now renders graphs https://arche.readthedocs.io/en/latest/, #41
 ### Removed